+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link href="./index.css" rel="stylesheet" />
-    <title>Simple Counter</title>
+    <title>Simba Counter</title>
     <script type="module" src="./index.js"></script>
   </head>
